% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSG2category}
\alias{ENSG2category}
\title{ENSG annotation with GO categories including GO.BP, GO.CC , GO.MF and
reactome}
\format{
A character with 1490269 rows and 4 variables:
\describe{
  \item{source}{Annotation source}
  \item{gene_id}{ENSG gene id}
  \item{Category}{CAtegory annotation}
  \item{Name}{ENSG name}
}
}
\source{
mRNAinteractomeHeLa
}
\usage{
ENSG2category
}
\value{
A character.
}
\description{
ENSG annotation with GO categories including GO.BP, GO.CC , GO.MF and
reactome
}
\keyword{datasets}
