% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Index}
\alias{Index}
\title{Index 4kmer length annotation}
\format{
An array with 1490269 entries:
}
\source{
RBPmapHeLa
}
\usage{
Index
}
\value{
An array .
}
\description{
Index 4kmer length annotation
}
\keyword{datasets}
