% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniProtFeatures}
\alias{miniProtFeatures}
\title{Protein feature annotation}
\format{
An list with 3 entries:
\describe{
\item{1}{AAStringSet of length 69025 with protein sequence information}
\item{2}{ENSG to gene annotation}
\item{3}{Gene name to emtry annotation}
}
}
\source{
RBDmap
}
\usage{
miniProtFeatures
}
\value{
A list.
}
\description{
Protein feature annotation
}
\keyword{datasets}
