% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QWCLpeptidesfiltered_clean}
\alias{QWCLpeptidesfiltered_clean}
\title{QWCLpeptidesfiltered_clean}
\usage{
data(QWCLpeptidesfiltered_clean)
}
\description{
QFeatures containing 1 assays with 24594 rows and 9 columns.
}
\references{
Garcia-Moreno et al. System-wide Profiling of RNA-Binding Proteins Uncovers Key Regulators of Virus Infection. Molecular Cell  2019 Apr 4;74:196-211.e11
}
\keyword{datasets}
