% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RICpeptides.raw}
\alias{RICpeptides.raw}
\title{RNA-bound proteomes of a SILAC experiment of SINV-infected cells
during a timecourse experiment.}
\format{
A dataframe with 6909 rows and 151 variables:
\describe{
  \item{Sequence}{The identified AA sequence of the peptide.}
  \item{Length}{The length of the sequence stored in the column 'Sequence'.}
  \item{Proteins}{Identifiers of proteins this site is associated with..}
  \item{“Intensity L ” + A/B/C exp}{Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the light label partner.}
  \item{“Intensity M ” + A/B/C exp}{Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the medium label partner.}
  \item{“Intensity H ” + A/B/C exp}{Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the heavy label partner.}
}
}
\source{
Garcia-Moreno et al 2019.
System-wide Profiling of RNA-Binding Proteins Uncovers Key Regulators of Virus Infection.
Molecular Cell  2019 Apr 4;74:196-211.e11
\href{https://pubmed.ncbi.nlm.nih.gov/30799147/}{10.1016/j.molcel.2019.01.017}.
}
\usage{
RICpeptides.raw
}
\value{
A data.frame.
}
\description{
The dataset contains the peptide.txt output file from
\href{http://www.maxquant.org}{MaxQuant}.
}
\keyword{datasets}
